+++ Tests/Unit/Stubs/Microsoft365.psm1
@@ -1,5 +1,250 @@
 
 #region ExchangeOnline
+function Get-AuthenticationPolicy
+{
+    [CmdletBinding()]
+    param(
+        [Parameter()]
+        [System.String]
+        $Identity,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthActiveSync,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthAutodiscover,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthImap,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthMapi,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthOfflineAddressBook,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthOutlookService,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthPop,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthPowerShell,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthReportingWebServices,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthRpc,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthSmtp,
+
+        [Parameter()]
+        [ValidateSet('Present', 'Absent')]
+        [System.String]
+        $Ensure,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthWebServices
+    )
+}
+function Set-AuthenticationPolicy
+{
+    [CmdletBinding()]
+    param(
+        [Parameter()]
+        [System.String]
+        $Identity,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $Confirm,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthActiveSync,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthAutodiscover,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthImap,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthMapi,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthOfflineAddressBook,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthOutlookService,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthPop,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthPowerShell,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthReportingWebServices,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthRpc,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthSmtp,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthWebServices
+    )
+}
+function New-AuthenticationPolicy
+{
+    [CmdletBinding()]
+    param(
+        [Parameter()]
+        [System.String]
+        $Name,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $Confirm,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthActiveSync,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthAutodiscover,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthImap,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthMapi,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthOfflineAddressBook,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthOutlookService,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthPop,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthPowerShell,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthReportingWebServices,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthRpc,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthSmtp,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthWebServices
+    )
+}
+function Remove-AuthenticationPolicy
+{
+    [CmdletBinding()]
+    param(
+        [Parameter()]
+        [System.String]
+        $Identity,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $Confirm,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthActiveSync,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthAutodiscover,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthImap,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthMapi,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthOfflineAddressBook,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthOutlookService,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthPop,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthPowerShell,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthReportingWebServices,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthRpc,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthSmtp,
+
+        [Parameter()]
+        [Switch]
+        $AllowBasicAuthWebServices
+    )
+}
 function Add-AvailabilityAddressSpace
 {
     [CmdletBinding()]
@@ -1110,6 +1355,15 @@ function Get-SupervisoryReviewRule
         [System.Object]
         $Policy,
 
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $AsJob
+    )
+}
+function Get-TransportConfig
+{
+    [CmdletBinding()]
+    param(
         [Parameter()]
         [System.Management.Automation.SwitchParameter]
         $AsJob
@@ -8875,6 +9129,103 @@ function Set-SharingPolicy
         [Parameter()]
         [System.Object]
         $Enabled,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $AsJob
+    )
+}
+function Set-TransportConfig
+{
+    [CmdletBinding()]
+    param(
+        [Parameter()]
+        [System.Object]
+        $AddressBookPolicyRoutingEnabled,
+
+        [Parameter()]
+        [System.Object]
+        $ClearCategories,
+
+        [Parameter()]
+        [System.Object]
+        $ConvertDisclaimerWrapperToEml,
+
+        [Parameter()]
+        [System.Object]
+        $DSNConversionMode,
+
+        [Parameter()]
+        [System.Object]
+        $ExternalDelayDsnEnabled,
+
+        [Parameter()]
+        [System.Object]
+        $ExternalDsnLanguageDetectionEnabled,
+
+        [Parameter()]
+        [System.Object]
+        $ExternalDsnSendHtml,
+
+        [Parameter()]
+        [System.Object]
+        $ExternalPostmasterAddress,
+
+        [Parameter()]
+        [System.Object]
+        $HeaderPromotionModeSetting,
+
+        [Parameter()]
+        [System.Object]
+        $InternalDelayDsnEnabled,
+
+        [Parameter()]
+        [System.Object]
+        $InternalDsnLanguageDetectionEnabled,
+
+        [Parameter()]
+        [System.Object]
+        $InternalDsnSendHtml,
+
+        [Parameter()]
+        [System.Object]
+        $JournalingReportNdrTo,
+
+        [Parameter()]
+        [System.Object]
+        $JournalMessageExpirationDays,
+
+        [Parameter()]
+        [System.Object]
+        $MaxRecipientEnvelopeLimit,
+
+        [Parameter()]
+        [System.Object]
+        $ReplyAllStormBlockDurationHours,
+
+        [Parameter()]
+        [System.Object]
+        $ReplyAllStormDetectionMinimumRecipients,
+
+        [Parameter()]
+        [System.Object]
+        $ReplyAllStormDetectionMinimumReplies,
+
+        [Parameter()]
+        [System.Object]
+        $ReplyAllStormProtectionEnabled,
+
+        [Parameter()]
+        [System.Object]
+        $Rfc2231EncodingEnabled,
+
+        [Parameter()]
+        [System.Object]
+        $SmtpClientAuthenticationDisabled,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $Confirm,
 
         [Parameter()]
         [System.Management.Automation.SwitchParameter]