+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change log for Microsoft365DSC
 
+# UNRELEASED
+
+* EXOAuthenticationPolicy
+  * Initial release;
+
+# 1.22.316.1
+
+* EXOCASMailboxPlan
+  * Add support for DisplayName as identifier for CAS mailbox plan.
+* EXOTransportSettings
+  * New resource for Exchange Online transport configuration.
+* IntuneAppProtectionPolicyiOS
+  * Add 7 additional parameters to the resource and added parameter
+    descriptions.
+* DEPENDENCIES
+  * Updated DSCParser to 1.3.0.4.
+  * Updated Microsoft.Graph.* to 1.9.3.
+
 # 1.22.309.1
 
 * EXOAcceptedDomain
@@ -1553,7 +1571,3 @@ MISC
 * SPOSite
   * Added default value for Storage Quota;
   * Fixed an issue with site creation that could result in infinite loops;
-
-## UNRELEASED
-* IntuneDeviceConfigurationPolicyAndroidWorkProfile
-  * Initial release;